update `meta/main.yml`
I was running into the same issue where the dependency I had in `meta/main.yml` was listed as "not a type of object". I was able to fix this on my role by prepending `name: ` to the dependency entry.  It was a bear to find this since the ansible documentation examples consistently don't show this.